diagnostics: Add the diagnostic API
The diagnostics in Wave take a lot of inspiration from elm and rust. They
contain a span, which indicates the source of the diagnostic, a
message, which explains the error, a label, which is a string
attached to the snippet of code, and a hint, which should explain what
to do to fix the error/warning.