parser: add parser.c and parser.h
These two file contain the implementation of the parser. The
implementation is not complete: only a little part of the language
is implemented.

test.c: edit the diagnostic test

The hint got extracted to a variable and it became similar to one of the
error that we produce in the parser. We changed the test to try to
figure out how to format the hints to be better and easier to comprehend.