parser: Add pop_node()
With this commit, we add a new function which pops the last node from
the NodeList of the parser. Before popping, it checks if what you
want to pop is the last element of the array. If this condition is
not met the compiler will panic with an Internal Compiler Error.